% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dviGLR.R
\name{dviGLR}
\alias{dviGLR}
\title{Finds Generalised Likelihood Ratios (GLRs)}
\usage{
dviGLR(
  dvi,
  pairings = generatePairings(dvi),
  disableMutations = F,
  verbose = F
)
}
\arguments{
\item{dvi}{A `dviData` object.}

\item{disableMutations}{A logical, or NA (default). The default action is to
disable mutations in all reference families without Mendelian errors.}

\item{verbose}{A logical.}

\item{list}{Hypotheses.}
}
\value{
A data frame with GLRs
}
\description{
Based on a `dviData` objects, the GLR for specified tests are computed
}
\examples{
dviGLR(example2, pairings = list(V1 = "M1"))

# All tests with output from dviJoint
dviGLR(example2, verbose = T)

}
