% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplePedigrees.R
\name{examplePedigrees}
\alias{examplePedigrees}
\title{Generates pedigrees for grandmother example}
\usage{
examplePedigrees(plot = T, pedNames = NULL, gm = T)
}
\arguments{
\item{plot}{logical}

\item{pedNames}{character vector names of pedigrees}

\item{gm}{logical}
}
\value{
A list of pedigrees where `ped0` corresponds to H0: `unrelated`.
}
\description{
Optionally plots
}
\details{
Two examples are implemented, the grandmoter example (`gm = T`)
or with one pedigree added (`gm = F`).
}
\examples{
# peds = examplePedigrees(pedNames = NULL)
peds = examplePedigrees(pedNames = paste0("Ped", 0:2))
#peds = examplePedigrees(gm = F)
#peds = examplePedigrees(gm = F, pedNames = paste0("Ped", 0:3))
}
