% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findLR.R
\name{findLR}
\alias{findLR}
\title{Finds LRs and the maximum}
\usage{
findLR(pedigrees, sim, truePed = NULL, ref = 1)
}
\arguments{
\item{pedigrees}{a list of pedigrees}

\item{sim}{pedigree with simulated marker data}

\item{truePed}{index or name of pedigree with marker data}

\item{ref}{index or name of pedigree in the numerator of LR}
}
\value{
A data frame with LRs, the max and log likelihood of the first pedigree,
the reference and optionally a plot.
}
\description{
n named pedigrees are given and simulated marker data. The LRs comparing
1 to n,..., (n-1) to n and the maximum Z = max(LR.1.n, ..., LR.n-1.n)
are calculated
}
\examples{
peds = examplePedigrees(plot = F, gm = F)
names(peds) = paste0("Ped", 0:3)
truePed  = "Ped0"
ref = "Ped1"
nsim = 2
conditional = F
peds[[truePed]] = setMarkers(peds[[truePed]],
                  locusAttributes = NorwegianFrequencies[1:22])
peds[[truePed]] = setMutmod(peds[[truePed]], model = "proportional", rate = 0.001)
sim = simPedigree(peds[[truePed]], idTarget ="2", idReferences = c("1", "3"),
                  nsim = nsim, seed = 1729, conditional = conditional)

res = findLR(peds, sim, truePed = truePed,  ref = ref)

}
