% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmPedigrees.R
\name{gmPedigrees}
\alias{gmPedigrees}
\title{Generates pedigrees for grandmother example}
\usage{
gmPedigrees(plot = T, tit = c("Ped1", "Ped2", "Ped3"))
}
\arguments{
\item{plot}{logical}

\item{tit}{title Character of length 3}
}
\value{
A list of pedigrees
}
\description{
Optionally plots
}
\examples{
peds = gmPedigrees()

}
