% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRmaxTest.R
\name{LRmaxTest}
\alias{LRmaxTest}
\title{Performs max LR test}
\usage{
LRmaxTest(pedigrees, sim, nsim = NULL, truePed = NULL)
}
\arguments{
\item{pedigrees}{a list of pedigrees}

\item{truePed}{integer, index of pedigree with marker data}
}
\value{
A data frame with LRs and the max
}
\description{
n named pedigrees are given, one with marker data. The LRs comparing
1 to n,..., (n-1) to n and the maximum Z = max(LR.1.n, ..., LR.n-1.n)
are calculated
}
\examples{
peds = gmPedigrees(plot = F)
truePed  = 2
nsim = 2
conditional = T
peds[[truePed]] = setMarkers(peds[[truePed]],
                  locusAttributes = NorwegianFrequencies[1:5])
peds[[truePed]] = setMutmod(peds[[truePed]], model = "proportional", rate = 0.001)
sim = simPedigree(peds[[truePed]], idTarget ="2", idReferences = c("1", "3"),
                  nsim = nsim, seed = 1729, conditional = conditional)

LRmaxTest(peds, sim, nsim = nsim, truePed = truePed)

}
