% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plots}
\alias{plots}
\title{Plot LRs and max}
\usage{
plots(res, plot = "box", ref = 1, truePed = 1, tit = NULL)
}
\arguments{
\item{res}{a data frame from`findLR()`}

\item{plot}{character, `box`, `density` or NULL}

\item{ref}{integer or character indicating true ped}
}
\value{
Various plots
}
\description{
Based on output from `findLR()` a boxplot of the LRs and Z or a density plot of
is produced
}
\examples{
peds = examplePedigrees(plot = F, gm = F)
names(peds) = paste0("Ped", 0:3)
truePed  = "Ped0"
ref = "Ped0"
peds[[truePed]] = setMarkers(peds[[truePed]],
                  locusAttributes = NorwegianFrequencies[1:22])
peds[[truePed]] = setMutmod(peds[[truePed]], model = "proportional", rate = 0.001)
sim = simPedigree(peds[[truePed]], idTarget ="2", idReferences = c("1", "3"),
                  nsim = 10, seed = 1729, conditional = F)

res = findLR(peds, sim,  truePed = truePed,  ref = ref)
plots(res, plot = "box", ref = ref)
plots(res, plot = "density", ref = ref)
}
