% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPedigree.R
\name{simPedigree}
\alias{simPedigree}
\title{Simulates marker data for a pedigree}
\usage{
simPedigree(
  ped,
  idTarget,
  idReferences,
  nsim = 1,
  seed = NULL,
  conditional = F
)
}
\arguments{
\item{ped}{a pedigree or list of pedigrees with markers attached}

\item{idTarget}{character POI}

\item{idReferences}{character vector References}

\item{integer}{seed}

\item{logic}{conditional}
}
\value{
Pedigree with simulations
}
\description{
Unconditional or conditional simulation.
}
\examples{
peds = gmPedigrees(plot = F)
truePed = 3
peds[[truePed]] <- setMarkers(peds[[truePed]],
             locusAttributes =  NorwegianFrequencies)
peds[[truePed]] <- setMutmod(peds[[truePed]], model = "proportional", rate = 0.001)
sim = simPedigree(peds[[truePed]], idTarget = "2", idReferences = c("2", "3"),
              nsim = 2, conditional = T)
sim

}
