% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmExample.R
\name{gmExample}
\alias{gmExample}
\title{Simulation for grandmother case}
\usage{
gmExample(pedigrees, Nsim, seed, conditional = F, simulateFrom = "Ped3")
}
\arguments{
\item{pedigrees}{a list of three pedigrees}

\item{Nsim}{integer}

\item{seed}{integer}

\item{conditional}{logical}

\item{character}{`Ped1`, `Ped2`, or `Ped3`}
}
\value{
A data frame with LR.1.3, LR.2.3 and max, i.e., LRmax
}
\description{
Conditional and unconditional simulation for hard coded example
}
\examples{
Set1 <- c(
  "CSF1PO", "D2S1338", "D3S1358", "D5S818", "D7S820",
  "D8S1179", "D13S317", "D16S539", "D18S51", "D19S433", "D21S11", "FGA",
  "TH01", "TPOX", "VWA"
)
db <- forrel::NorwegianFrequencies[Set1]
nsim = 2
seed17 = 17
cond = F
mut = T

# Generate pedigrees
peds = examplePedigrees(plot = F)

# Simulate from Ped0
peds[["Ped0"]] <- setMarkers(peds[["Ped0"]], locusAttributes = db)
if(mut) peds[["Ped0"]] <- setMutmod(peds[["Ped0"]], model = "proportional",
rate = 0.001)
LR1 <- gmExample(
  pedigrees = peds, Nsim = nsim, seed = seed17,
  conditional = cond, simulateFrom = "Ped0"
)
log1Max <- log10(LR1[, 3])

# Simulate from Ped2
peds[["Ped1"]] <- setMarkers(peds[["Ped1"]], locusAttributes = db)
if(mut) peds[["Ped1"]] <- setMutmod(peds[["Ped1"]], model = "proportional", rate = 0.001)
LR2 <- gmExample(
  pedigrees = peds, Nsim = nsim, seed = seed17,
  conditional = cond, simulateFrom = "Ped1"
)
log2Max <- log10(LR2[, 3])

# Simulate from Ped3
peds[["Ped2"]] <- setMarkers(peds[["Ped2"]], locusAttributes = db)
if (mut) peds[["Ped2"]] <- setMutmod(peds[["Ped2"]], model = "proportional", rate = 0.001)
LR3 <- gmExample(
  pedigrees = peds, Nsim = nsim, seed = seed17,
  conditional = cond, simulateFrom = "Ped2"
)
log3Max <- log10(LR3[, 3])
boxplot(log1Max, log2Max, log3Max, names = c("Ped0", "Ped1", "Ped2"),
         ylab = "log10(LRmax)")
criticalValue = quantile(log1Max, probs = 0.95)
abline(h = criticalValue)
}
